+++ tests/testthat/test_h_hdbscan.R
@@ -108,6 +108,14 @@ test_that("failing example doesn't fail", {
 	expect_silent(hdbscanobj <- hdbscan(vis, minPts = 10, K = 5, threads = 2))
 })
 
+test_that("glosh is in range", {
+	data(iris)
+	expect_silent(vis <- largeVis(t(iris[,1:4]), K = 20, sgd_batches = 1, threads = 2))
+	expect_silent(hdbscanobj <- hdbscan(vis, minPts = 10, K = 5, threads = 2))
+	expect_equal(sum(hdbscanobj$glosh < 0), 0)
+	expect_equal(sum(hdbscanobj$glosh > 1), 0)
+})
+
 context("gplot")
 
 set.seed(1974)