+++ NEWS.md
@@ -1,5 +1,7 @@
 ### largeVis 0.1.11
 * Performance improvements in neighbor search & projectKNNs.
+* Performance improvements in HDBSCAN. 
+* HDBSCAN now outputs GLOSH outlier scores for each point assigned to a cluster.
 
 ### largeVis 0.1.10.1
 * Hotfix for issue the caused largeVis to fail if compiled without 64-bit ARMA